+++ portal/handlers.py
@@ -40,7 +40,7 @@ from django.db.models.signals import post_save, pre_delete
 from django.dispatch import receiver
 from django_otp.models import Device
 
-from portal.utils import two_factor_cache_key
+from common.utils import two_factor_cache_key
 
 
 @receiver([post_save, pre_delete])
