+++ portal/tests/test_independent_student.py
@@ -41,20 +41,20 @@ import time
 from django.core import mail
 from selenium.webdriver.support.wait import WebDriverWait
 
+from common.tests.utils import email as email_utils
 from common.tests.utils.classes import create_class_directly
-from common.tests.utils.teacher import signup_teacher_directly
-from common.utils.student import (
+from common.tests.utils.student import (
     create_independent_student,
     create_independent_student_directly,
+    create_school_student_directly,
     signup_duplicate_independent_student_fail,
     submit_independent_student_signup_form,
 )
+from common.tests.utils.teacher import signup_teacher_directly
 from portal.tests.utils.organisation import create_organisation_directly
-from portal.testscommon.utils.student import create_school_student_directly
 
 from .base_test import BaseTest
 from .pageObjects.portal.home_page import HomePage
-from .utils import email as email_utils
 from .utils.messages import (
     is_email_verified_message_showing,
     is_indep_student_join_request_received_message_showing,
