+++ portal/tests/test_migration_preview_user_remove.py
@@ -36,7 +36,7 @@
 # identified as the original program.
 from __future__ import absolute_import
 
-from portal.utils import field_exists
+from common.utils import field_exists
 from .base_test_migration import MigrationTestCase
