+++ portal/tests/test_organisation.py
@@ -40,14 +40,14 @@ from builtins import range
 
 from django.core import mail
 
+from common.tests.utils import email as email_utils
 from common.tests.utils.classes import create_class_directly
-from common.utils.student import create_school_student_directly
-from common.utils.teacher import generate_details, signup_teacher_directly
+from common.tests.utils.student import create_school_student_directly
+from common.tests.utils.teacher import generate_details, signup_teacher_directly
 from portal.tests.pageObjects.portal.base_page import BasePage
 from portal.tests.pageObjects.portal.home_page import HomePage
 
 from .base_test import BaseTest
-from .utils import email as email_utils
 from .utils.messages import is_organisation_created_message_showing
 from .utils.organisation import (
     create_organisation,
