+++ portal/tests/test_school_student.py
@@ -37,8 +37,8 @@
 from __future__ import absolute_import
 
 from common.tests.utils.classes import create_class_directly
-from common.utils.student import create_school_student_directly
-from common.utils.teacher import signup_teacher_directly
+from common.tests.utils.student import create_school_student_directly
+from common.tests.utils.teacher import signup_teacher_directly
 from portal.tests.pageObjects.portal.home_page import HomePage
 
 from .base_test import BaseTest
