+++ portal/tests/test_teacher_student.py
@@ -39,12 +39,12 @@ from __future__ import absolute_import
 import time
 
 from common.tests.utils.classes import create_class_directly
-from common.utils.student import (
+from common.tests.utils.student import (
     create_many_school_students,
     create_school_student,
     create_school_student_directly,
 )
-from common.utils.teacher import signup_teacher_directly
+from common.tests.utils.teacher import signup_teacher_directly
 from portal.tests.pageObjects.portal.home_page import HomePage
 
 from .base_test import BaseTest
