+++ portal/views/organisation.py
@@ -46,7 +46,7 @@ from django.shortcuts import render, get_object_or_404
 from rest_framework.authentication import SessionAuthentication
 from rest_framework.views import APIView
 
-import portal.permissions as permissions
+import common.permissions as permissions
 from portal import email_messages
 from portal.forms.organisation import OrganisationJoinForm, OrganisationForm
 from common.helpers.emails import send_email, NOTIFICATION_EMAIL
