+++ portal/views/registration.py
@@ -62,7 +62,7 @@ from portal.forms.registration import (
 )
 from portal.helpers.captcha import remove_captcha_from_form
 from common.helpers.emails import PASSWORD_RESET_EMAIL
-from portal.permissions import not_logged_in, not_fully_logged_in
+from common.permissions import not_logged_in, not_fully_logged_in
 
 
 @user_passes_test(not_logged_in, login_url=reverse_lazy("home"))
