+++ portal/views/teacher/solutions_level_selector.py
@@ -43,7 +43,7 @@ from django.utils.safestring import mark_safe
 from game.models import Episode
 from django.core.cache import cache
 from game import app_settings
-from portal.permissions import logged_in_as_teacher
+from common.permissions import logged_in_as_teacher
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.core.urlresolvers import reverse_lazy
 from django.shortcuts import render, get_object_or_404
